+++ api/http_api.go
@@ -1,4 +1,4 @@
-package http
+package api
 
 import (
 	"fmt"
